Hexo template
Since the Travis build is already set up for npm for the firebase tools,
this PR uses Hexo, which seems like a relatively popular Octopress-alike
written in JavaScript:

https://hexo.io/

I've never used this before and I'm not exactly a huge fan of
JavaScript, but this seems relatively popular and probably fine.

The main requirement is Markdown support. This has Markdown support!