Move `src` to `site`.
This allows using `src` for a Rust crate.

Why would we want that? Because `mdbook` does not yet have support for
expressing dependencies. Therefore, to make rustdoc tests that require
`extern crate` declarations, you need to make a stub crate that depends
on the external crates, then manually pass linker arguments to rustdoc.